#732 Postgresql changes
